embedded-usb-pd: Audit panics (#60)
This PR audits potential panics. Most were cases that are easily proven
to never panic (assuming the assumptions in the code are held) and these
assumptions are defined very close to the potential panic path, so I
think it was safe to allow these instead of returning an error.

A refactor of the `Args` type may help eliminate the need for unwraps
(see:
https://github.com/OpenDevicePartnership/embedded-usb-pd/issues/61).

`Rdo::for_pdo` now returns an `Option<Self>` and is public so a minor
breaking change has been introduced.

Resolves #59